+++ src/App.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import './App.scss';
-import { GoodsList } from './GoodsList';
-
+import { GoodsList } from './components/GoodsList';
 import { getAll, get5First, getRedGoods } from './api/goods';
 import { Good } from './types/Good';
+import 'bulma';
 
 export const App: React.FC = () => {
   const [goods, setGoods] = useState<Good[] | []>([]);
@@ -16,33 +16,40 @@ export const App: React.FC = () => {
 
   return (
     <div className="App">
-      <h1>Dynamic list of Goods</h1>
+      <h1 className="title has-text-centered">
+        Dynamic list of Goods
+      </h1>
 
-      <button
-        type="button"
-        data-cy="all-button"
-        onClick={() => loadUsers(getAll)}
-      >
-        Load all goods
-      </button>
+      <div className="container has-text-centered">
+        <button
+          type="button"
+          className="button is-success block mr-5"
+          data-cy="all-button"
+          onClick={() => loadUsers(getAll)}
+        >
+          Load all goods
+        </button>
 
-      <button
-        type="button"
-        data-cy="first-five-button"
-        onClick={() => loadUsers(get5First)}
-      >
-        Load 5 first goods
-      </button>
+        <button
+          type="button"
+          className="button is-warning block"
+          data-cy="first-five-button"
+          onClick={() => loadUsers(get5First)}
+        >
+          Load 5 first goods
+        </button>
 
-      <button
-        type="button"
-        data-cy="red-button"
-        onClick={() => loadUsers(getRedGoods)}
-      >
-        Load red goods
-      </button>
+        <button
+          type="button"
+          className="button is-danger block ml-5"
+          data-cy="red-button"
+          onClick={() => loadUsers(getRedGoods)}
+        >
+          Load red goods
+        </button>
 
-      <GoodsList goods={goods} />
+        <GoodsList goods={goods} />
+      </div>
     </div>
   );
 };
