+++ src/components/GoodsList/index.ts
@@ -1,0 +1,1 @@
+export * from './GoodsList';
